ref(ct): fixes according to code review
Chages:
- move `BaseForm` from `ct/admin.py` to `ct/forms.py`
- chage models variable to `MODELS` module level variable
- change import statements from `import ct.models` to `from .models import (...)`
